fix: guard navigator.clipboard before writeText and handle write rejections
before: copying in SSR, non-HTTPS contexts, or older browsers without the Clipboard API threw a ReferenceError, and permission-denied rejections surfaced as unhandled promise rejections.

after:
- `useActionBarCopy` in `@assistant-ui/core` wraps the optional `copyToClipboard` call in `.then(success, () => {})` so rejections are absorbed silently.
- `ActionBarPrimitive.Copy` in `@assistant-ui/react` returns a rejected promise when the Clipboard API is unavailable, which then flows through the core rejection handler.
- `packages/ui` shadcn copy-target `markdown-text.tsx` adds the same guard inline (kept self-contained for shadcn copy semantics).
- All 6 templates apply the same inline guard.
- A new shared hook `@assistant-ui/ui/hooks/use-copy-to-clipboard` carries the guarded implementation; `apps/docs` markdown viewer imports it instead of keeping its own duplicate.

based on the original fix from #3513 by @JustAnOkapi.
