+++ templates/cloud/components/assistant-ui/markdown-text.tsx
@@ -55,12 +55,17 @@ const useCopyToClipboard = ({
   const [isCopied, setIsCopied] = useState<boolean>(false);
 
   const copyToClipboard = (value: string) => {
-    if (!value) return;
+    if (!value || typeof navigator === "undefined" || !navigator.clipboard) {
+      return;
+    }
 
-    navigator.clipboard.writeText(value).then(() => {
-      setIsCopied(true);
-      setTimeout(() => setIsCopied(false), copiedDuration);
-    });
+    navigator.clipboard.writeText(value).then(
+      () => {
+        setIsCopied(true);
+        setTimeout(() => setIsCopied(false), copiedDuration);
+      },
+      () => {},
+    );
   };
 
   return { isCopied, copyToClipboard };
